feat:extract stopwords module
